+++ posthog/src/main/java/com/posthog/internal/PostHogDeviceDateProvider.kt
@@ -1,5 +1,6 @@
 package com.posthog.internal
 
+import com.posthog.PostHogInternal
 import java.util.Calendar
 import java.util.Date
 
@@ -7,7 +8,8 @@ import java.util.Date
  * The implementation of the PostHogDateProvider interface that reads the current time
  *  and more based on the device's clock.
  */
-internal class PostHogDeviceDateProvider : PostHogDateProvider {
+@PostHogInternal
+public class PostHogDeviceDateProvider : PostHogDateProvider {
     override fun currentDate(): Date {
         val cal = Calendar.getInstance()
         return cal.time