+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+- feat: use network time instead of device time ([#235](https://github.com/PostHog/posthog-android/pull/235))
+
 ## 3.13.1 - 2025-04-01
 
 - fix: add remote config response to proguard rules ([#238](https://github.com/PostHog/posthog-android/pull/238))